app/requirements.txt
Django
djangorestframework
psycopg2
celery
redis
requests
asyncio
aiohttp